split up file into an entire package
