+++ airbyte-integrations/connectors/destination-dev-null/src/test-integration/kotlin/io/airbyte/integrations/destination/dev_null/DevNullCheckIntegrationTest.kt
@@ -6,7 +6,7 @@ package io.airbyte.integrations.destination.dev_null
 
 import io.airbyte.cdk.test.check.CheckIntegrationTest
 import io.airbyte.cdk.test.check.CheckTestConfig
-import io.airbyte.cdk.test.util.TestDeploymentMode
+import io.airbyte.cdk.test.util.destination_process.TestDeploymentMode
 import java.util.regex.Pattern
 import org.junit.jupiter.api.Test
